+++ tests/_test.py
@@ -4,27 +4,7 @@ tests._test
 """
 from object_colors import Color
 
-from . import (
-    ATTRS,
-    CODES,
-    COLORED_DUPES,
-    DUPE_MARKED,
-    DUPE_MARKED_COLOR,
-    DUPES,
-    EXACT_INDEX_COLOR,
-    GREEN,
-    INSTANCES,
-    LONG_STRING,
-    MARKED,
-    MARKED_COLOR,
-    MARKED_SECOND,
-    MARKED_SECOND_COLOR,
-    RED,
-    RESET,
-    SPACED_WORDS,
-    SPACED_WORDS_COLOR,
-    TEST_STR,
-)
+from . import ATTRS, GREEN, INSTANCES, RED, RESET, TEST_STR
 
 
 def test_color_string():
@@ -60,147 +40,25 @@ def test_str_args():
     """Test for the correct ANSI codes for color, effect, and
     background.
     """
-    _color = Color("red", "bold", "green")
-    assert _color.fore == 1
-    assert _color.effect == 1
-    assert _color.back == 2
+    color = Color("red", "bold", "green")
+    assert color.fore == 1
+    assert color.effect == 1
+    assert color.back == 2
 
 
 def test_str_ints():
     """Test the correct ANSI code integers are produced."""
-    _color = Color(1, 1, 2)
-    assert _color.fore == 1
-    assert _color.effect == 1
-    assert _color.back == 2
+    color = Color(1, 1, 2)
+    assert color.fore == 1
+    assert color.effect == 1
+    assert color.back == 2
 
 
 def test_int_dict():
     """Test for a non-existing color and assert it will default to
     white instead of raising an error."""
-    _color = Color(orange=1)
-    assert _color.fore == 7
-
-
-def test_exact_letter_in_colored_string(populated_colors):
-    """For multicolored string assert the exact letter is the correct
-    color.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "c")
-    assert keys == TEST_STR
-    colored_keys = populated_colors.red.get_key(color_str, "c")
-    assert colored_keys == color_str
-
-
-def test_exact_word_in_string(populated_colors):
-    """Test the exact word selected to color is the appropriate
-    color.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "Cc:")
-    assert keys == MARKED
-    colored_keys = populated_colors.red.get_key(color_str, "Cc:")
-    assert colored_keys == MARKED_COLOR
-
-
-def test_exact_second_word_in_string(populated_colors):
-    """verify indices are working (at least for some)
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "My")
-    assert keys == MARKED_SECOND
-    colored_keys = populated_colors.red.get_key(color_str, "My")
-    assert colored_keys == MARKED_SECOND_COLOR
-
-
-def test_dupe_words(populated_colors):
-    """Test the correct output is produced when there are duplicate
-    words in the string.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    keys = populated_colors.red.get_key(DUPES, "one")
-    assert keys == DUPE_MARKED
-    colored_keys = populated_colors.red.get_key(COLORED_DUPES, "one")
-    assert colored_keys == DUPE_MARKED_COLOR
-
-
-def test_word_in_string(populated_colors):
-    """Test a colored word exists in string.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "cc:")
-    assert keys == TEST_STR
-    colored_keys = populated_colors.red.get_key(color_str, "cc:")
-    assert colored_keys == color_str
-
-
-def test_words_in_string(populated_colors):
-    """Test colored words exist in string.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "Cc:", "Business")
-    assert keys == SPACED_WORDS
-    colored_keys = populated_colors.red.get_key(color_str, "Cc:", "Business")
-    assert colored_keys == SPACED_WORDS_COLOR
-
-
-def test_exact_letter_in_string(populated_colors):
-    """Test exact letter exists in string.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    colored_keys = populated_colors.red.get_key(
-        color_str, "c", ignore_case=True, scatter=True
-    )
-    assert colored_keys == EXACT_INDEX_COLOR
-
-
-def test_exact_word_in_string_ignore_case(populated_colors):
-    """Test word is in string, whether letters are a mixture of upper
-    and lower case, or not.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "Cc:", ignore_case=True)
-    assert keys == MARKED
-    colored_keys = populated_colors.red.get_key(
-        color_str, "Cc:", ignore_case=True
-    )
-    assert colored_keys == MARKED_COLOR
-
-
-def test_word_in_string_ignore_case(populated_colors):
-    """Test word is in string, whether letters are a mixture of upper
-    and lower case, or not.
-    """
-    color_str = populated_colors.green.get(TEST_STR)
-    keys = populated_colors.red.get_key(TEST_STR, "cc:", ignore_case=True)
-    assert keys == MARKED
-    colored_keys = populated_colors.red.get_key(
-        color_str, "cc:", ignore_case=True
-    )
-    assert colored_keys == MARKED_COLOR
+    color = Color(orange=1)
+    assert color.fore == 7
 
 
 def test_tuple_return(populated_colors):
@@ -218,41 +76,6 @@ def test_tuple_return(populated_colors):
     )
 
 
-def test_multi_values(populated_colors):
-    """Test multiple ANSI and color values at once.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    multi = populated_colors.multicolor(LONG_STRING)
-    ansis = populated_colors.get_list(multi)
-    for ansi in ansis:
-        if Color.ansi_escape.match(ansi):
-            for string in ansi:
-                if string.isdigit():
-                    assert int(string) <= 7
-
-
-def test_none():
-    """Test the correct string is returned when no keyword is used for
-    multicolor.
-    """
-    _color = Color(fore="GREEN")
-    none = _color.multicolor(LONG_STRING)
-    assert none == LONG_STRING
-
-
-def test_with_populated_colors(populated_colors):
-    """Test all the colors can be populated at once.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    """
-    all_colors = populated_colors.multicolor(LONG_STRING)
-    for code in CODES:
-        assert code in all_colors
-
-
 def test_pop_result(populated_colors):
     """Test colors can be removed from the instance.
 
@@ -262,12 +85,15 @@ def test_pop_result(populated_colors):
     assert hasattr(populated_colors, "red")
     red = populated_colors.pop("red")
     assert "red" not in populated_colors.__dict__
-    red_string = red.get("This is red")
-    assert red_string == "\u001b[0;31;40mThis is red\u001b[0;0m"
+    red_string = red.get(TEST_STR)
+    assert red_string == f"\u001b[0;31;40m{TEST_STR}\u001b[0;0m"
 
 
 def test_pop_no_result(color):
-    """Test that error goes by silently if item does not exist."""
+    """Test that error goes by silently if item does not exist.
+
+    :param color: Instantiated ``Color`` object.
+    """
     assert "red" not in color.__dict__
     red = color.pop("red")
     assert red is None
@@ -281,19 +107,6 @@ def test_color_print(capsys, populated_colors):
     :param populated_colors:    Instantiated ``Color`` object where
                                 ``populate_colors`` has been called.
     """
-    populated_colors.red.print("This stdout is red")
+    populated_colors.red.print(TEST_STR)
     captured = capsys.readouterr()
-    assert captured.out == f"{RED}This stdout is red{RESET}\n"
-
-
-def test_print_multi(capsys, populated_colors):
-    """Test printing with multi-colored string.
-
-    :param populated_colors:    Instantiated ``Color`` object where
-                                ``populate_colors`` has been called.
-    :param capsys:              ``pytest`` fixture to capture output.
-    """
-    populated_colors.print(LONG_STRING, multi=True)
-    captured = capsys.readouterr()
-    for code in CODES:
-        assert code in captured.out
+    assert captured.out == f"{RED}{TEST_STR}{RESET}\n"