+++ whitelist.py
@@ -1,4 +1,3 @@
 fixture_color  # unused function (tests/conftest.py:9)
 fixture_populated_colors  # unused function (tests/conftest.py:15)
-_.populate_colors  # unused method (object_colors/__init__.py:484)
-_.get_key  # unused method (object_colors/__init__.py:514)
+_.populate_colors  # unused method (object_colors/__init__.py:206)